+++ src/components/LayerMapsforge.tsx
@@ -38,6 +38,8 @@ export type LayerMapsforgeProps = {
 	renderTheme?: `/${string}` | typeof BUILT_IN_THEMES[number];
 	renderStyle?: string;
 	renderOverlays?: string[];
+	enabledZoomMin?: number;
+	enabledZoomMax?: number;
 	onRemove?: null | ( ( response: ResponseBase ) => void );
 	onCreate?: null | ( ( response: LayerMapsforgeResponse ) => void );
 	onChange?: null | ( ( response: LayerMapsforgeResponse ) => void );
@@ -51,6 +53,8 @@ const LayerMapsforge = ( {
 	renderTheme = 'DEFAULT',
 	renderStyle = '',
 	renderOverlays = [],
+    enabledZoomMin = 1,
+    enabledZoomMax = 30,
 	onCreate,
 	onRemove,
 	onChange,
@@ -79,6 +83,8 @@ const LayerMapsforge = ( {
 				renderTheme,
 				renderStyle,
 				renderOverlays,
+				Math.round( enabledZoomMin ),
+				Math.round( enabledZoomMax ),
 				reactTreeIndex
 			).then( ( response : LayerMapsforgeResponse ) => {
 				setUuid( response.uuid );
@@ -111,6 +117,17 @@ const LayerMapsforge = ( {
 		nativeNodeHandle,
 		!! uuid,
 		triggerCreateNew,
+	] );
+
+	// enabledZoomMin enabledZoomMax changed.
+	useEffect( () => {
+		if ( nativeNodeHandle && uuid ) {
+			Module.updateEnabledZoomMinMax( nativeNodeHandle, uuid, Math.round( enabledZoomMin ), Math.round( enabledZoomMax ) )
+			.catch( ( err: any ) => { console.log( 'ERROR', err ); onError ? onError( err ) : null } );
+		}
+	}, [
+		enabledZoomMin,
+		enabledZoomMax,
 	] );
 
 	useEffect( () => {