+++ src/compose/useMapEvents.ts
@@ -8,19 +8,19 @@ import { NativeEventEmitter } from 'react-native';
 /**
  * Internal dependencies
  */
-import type { mapEvent  } from '../types';
+import type { MapEventResponse  } from '../types';
 
 const useMapEvents = ( {
 	nativeNodeHandle,
 	onMapEvent,
 } : {
 	nativeNodeHandle: null | number,
-	onMapEvent?: null | ( ( response: mapEvent ) => void ),
+	onMapEvent?: null | ( ( response: MapEventResponse ) => void ),
 } ) : void => {
 
 	useEffect( () => {
 		const eventEmitter = new NativeEventEmitter();
-		let eventListener = eventEmitter.addListener( 'onMapEvent', response => {
+		let eventListener = eventEmitter.addListener( 'onMapEvent', ( response : MapEventResponse ) => {
 			if ( response.nativeNodeHandle === nativeNodeHandle && isFunction( onMapEvent ) ) {
                 onMapEvent( response );
 			}